requirements.txt
python-telegram-bot
pyautogui
requests
